refactor: v0.3.8 enum class 化 (ESI/ESIFC/AllocKind/IoDir + MapIRQHook + FCEU_ENUM_CLASS_BITFLAGS)
Plan v3 §5 v0.3.8 C-track continuation. Five categories of pre-v0.3.x
C-style typing replaced by scoped enumerations under fceu11::; legacy
spellings preserved as global using/inline constexpr aliases so the
~200 pre-existing call sites compile unchanged.

Conversions:
1. ESI / ESIFC  (src/git.h)            → enum class fceu11::InputDevice  : int8_t
                                       → enum class fceu11::InputDeviceFC: int8_t
2. fceuAllocType (src/utils/scoped_ptr.h) → enum class fceu11::AllocKind  : uint8_t
3. FCEUIOD_*    (src/driver.h)         → enum class fceu11::IoDir        : uint8_t
4. MapIRQHook   (src/x6502.h)          → using fceu11::MapIRQHook = void(*)(int)
                                       (global linkage preserved; ~35 mapper sites unchanged)
5. FCEU_ENUM_CLASS_BITFLAGS(E) macro   → src/utils/enum_class_bitflags.h
                                       + showcase enum class fceu11::CpuFlag (not in x6502.cpp hot path)

Plan deviations (with rationale in CHANGELOG.md):
- `: int8_t` instead of `: u8` for InputDevice/InputDeviceFC — SI_UNSET=-1
  sentinel preserved per iron-rule 1 byte-level consistency.
- Actual symbol is `fceuAllocType` (post v0.3.7), not `FCEU_ALLOC_TYPE` as
  plan §5 wrote pre-v0.3.7 split.
- "175 mapper" → measured 35 files / 42 sites (v0.3.4 dead-code cleanup
  reduced base).
- OldFceuApi = fceu11::v0_2_compat shim DEFERRED to v0.3.10 per §6.1
  phase-1 ("only new symbols enter fceu11::; old symbols stay global").
  v0.3.8 has no payload for the sub-namespace — every legacy alias is
  already at file scope.

API boundary changes (where ESI/ESIFC type appears in storage/struct):
- FCEUGI::input[2] / inputfc — typed; ROM-detection assignments in
  fceu.cpp/nsf.cpp/unif.cpp/vsuni.cpp use static_cast<ESI/ESIFC>(SI_*).
- FCEUI_SetInput[FC] signatures unchanged (already typed); 15 test call
  sites add static_cast<ESI/ESIFC> at the boundary.
- JOYPORT/FCPORT typed fields — JOYPORT ctor casts SI_UNSET.
- Qt InputConf.cpp: 17 QComboBox::addItem(text, SI_*/SIFC_*) gain (int)
  casts — Qt 6 QVariant rejects enum class implicit conversion.
- 2 switch(ESI_value) sites in input.cpp wrap operand in static_cast<int>.

Verification (five-gate plan §7):
- Gate 1 (compile): 0 errors, all targets built incl. new test exe.
- Gate 2 (unit): 6/6 ctest pass (smoke, mapper_load, mapper_reset,
  rom_regression, expected_api, NEW enum_class_bitflags).
- Gate 3 (byte-level): rom_regression_test 5 ROM SHA-256 match v0.3.0
  baseline (FCEUGI not in any SFORMAT mapping).
- Gate 4 (smoke): implicit via rom_regression_test (nestest path).
- cargo test: non-lua crates pass; fceux11-lua link issue is
  pre-existing (memory/test-runs-and-pitfalls.md).

Files: 26 changed (4 new, 22 modified). +471 −139 lines.

+++ src/drivers/Qt/InputConf.cpp
@@ -191,12 +191,17 @@ InputConfDialog_t::InputConfDialog_t(QWidget *parent)
 	for (int i = 0; i < 2; i++)
 	{
 		getInputSelection(i, &curNesInput[i], &usrNesInput[i]);
-		nesPortComboxBox[i]->addItem(tr("<None>"), SI_NONE);
-		nesPortComboxBox[i]->addItem(tr("Gamepad"), SI_GAMEPAD);
-		nesPortComboxBox[i]->addItem(tr("Zapper"), SI_ZAPPER);
-		nesPortComboxBox[i]->addItem(tr("Power Pad A"), SI_POWERPADA);
-		nesPortComboxBox[i]->addItem(tr("Power Pad B"), SI_POWERPADB);
-		nesPortComboxBox[i]->addItem(tr("Arkanoid Paddle"), SI_ARKANOID);
+		// v0.3.8: QVariant rejects enum class implicit conversion;
+		// SI_* are now `inline constexpr ESI` (= fceu11::InputDevice).
+		// Cast to int so QComboBox::addItem(text, userData) takes the
+		// numeric tag as before. itemData(j).toInt() consumers below
+		// stay unchanged.
+		nesPortComboxBox[i]->addItem(tr("<None>"), (int)SI_NONE);
+		nesPortComboxBox[i]->addItem(tr("Gamepad"), (int)SI_GAMEPAD);
+		nesPortComboxBox[i]->addItem(tr("Zapper"), (int)SI_ZAPPER);
+		nesPortComboxBox[i]->addItem(tr("Power Pad A"), (int)SI_POWERPADA);
+		nesPortComboxBox[i]->addItem(tr("Power Pad B"), (int)SI_POWERPADB);
+		nesPortComboxBox[i]->addItem(tr("Arkanoid Paddle"), (int)SI_ARKANOID);
 
 		for (int j = 0; j < nesPortComboxBox[i]->count(); j++)
 		{
@@ -212,17 +217,19 @@ InputConfDialog_t::InputConfDialog_t(QWidget *parent)
 	}
 
 	getInputSelection(2, &curNesInput[2], &usrNesInput[2]);
-	expPortComboxBox->addItem(tr("<None>"), SIFC_NONE);
-	expPortComboxBox->addItem(tr("Arkanoid Paddle"), SIFC_ARKANOID);
-	expPortComboxBox->addItem(tr("Shadow"), SIFC_SHADOW);
-	expPortComboxBox->addItem(tr("Hyper Shot Gun"), SIFC_HYPERSHOT);
-	expPortComboxBox->addItem(tr("Family Keyboard"), SIFC_FKB);
-	expPortComboxBox->addItem(tr("Mahjong"), SIFC_MAHJONG);
-	expPortComboxBox->addItem(tr("Quiz King Buzzers"), SIFC_QUIZKING);
-	expPortComboxBox->addItem(tr("Family Trainer A"), SIFC_FTRAINERA);
-	expPortComboxBox->addItem(tr("Family Trainer B"), SIFC_FTRAINERB);
-	expPortComboxBox->addItem(tr("Oeka Kids Tablet"), SIFC_OEKAKIDS);
-	expPortComboxBox->addItem(tr("Top Rider"), SIFC_TOPRIDER);
+	// v0.3.8: SIFC_* are now `inline constexpr ESIFC` (enum class) —
+	// cast to int for QVariant. See nesPortComboxBox block above.
+	expPortComboxBox->addItem(tr("<None>"), (int)SIFC_NONE);
+	expPortComboxBox->addItem(tr("Arkanoid Paddle"), (int)SIFC_ARKANOID);
+	expPortComboxBox->addItem(tr("Shadow"), (int)SIFC_SHADOW);
+	expPortComboxBox->addItem(tr("Hyper Shot Gun"), (int)SIFC_HYPERSHOT);
+	expPortComboxBox->addItem(tr("Family Keyboard"), (int)SIFC_FKB);
+	expPortComboxBox->addItem(tr("Mahjong"), (int)SIFC_MAHJONG);
+	expPortComboxBox->addItem(tr("Quiz King Buzzers"), (int)SIFC_QUIZKING);
+	expPortComboxBox->addItem(tr("Family Trainer A"), (int)SIFC_FTRAINERA);
+	expPortComboxBox->addItem(tr("Family Trainer B"), (int)SIFC_FTRAINERB);
+	expPortComboxBox->addItem(tr("Oeka Kids Tablet"), (int)SIFC_OEKAKIDS);
+	expPortComboxBox->addItem(tr("Top Rider"), (int)SIFC_TOPRIDER);
 
 	for (int j = 0; j < expPortComboxBox->count(); j++)
 	{
